% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGen.R
\name{MatrixGen}
\alias{MatrixGen}
\title{MatrixGen}
\usage{
MatrixGen(dataframe, clusters, amplicons)
}
\arguments{
\item{dataframe}{a dataframe cointains each cells in the row and each amplicon in the columns output of mosaic software}

\item{clusters}{a vector containing the name of the clusters to be analyzed}

\item{amplicons}{a vector containing the name of the amplicons to be analyzed}
}
\value{
matrix
}
\description{
generate a Matrix with the median ploidy for each amplicon suitable for sample-to-sample heatmap and differentially ploidy analysis
}
\examples{
#do not run
#load data
df <- read.table("/path-to-ploidy-table/ploidy.csv", sep = ",", header = TRUE)
clusters <- unique(df$label) #take all label 
amplicons <- colnames(df)[-c(1,2)]
TACOS::MatrixGen(df,clusters,amplicons)
}
