% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAmplicon.R
\name{plotAmplicon}
\alias{plotAmplicon}
\title{plotAmplicon}
\usage{
plotAmplicon(
  ploidy_table,
  amplicon,
  clusters,
  specific_clusters = FALSE,
  jitter = TRUE
)
}
\description{
Plot ploidy distribution of specific amplicon
}
