% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelevantClones.R
\name{RelevantClones}
\alias{RelevantClones}
\title{RelevantClones}
\usage{
RelevantClones(table)
}
\value{
a data.frame with phase information
}
\description{
Find clones with an increased frequency in phase2
}
\examples{
#do not run
TACOS::RelevandClones(summaries=c(exp1,exp2), fraction=.001, out=outhtml)
}
