% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffPloidy.R
\name{DiffPloidy}
\alias{DiffPloidy}
\title{DiffPloidy}
\usage{
DiffPloidy(
  ploidy_table,
  p1_cluster,
  p2_cluster,
  amplicons,
  method = c("wilcoxon", "kruskal", "t-test")
)
}
\description{
Differential ploidy analysis
}
