% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGen.R
\name{MatrixGen}
\alias{MatrixGen}
\title{MatrixGen}
\usage{
MatrixGen(dataframe, clusters, amplicons)
}
\arguments{
\item{dataframe}{a dataframe cointains each cells in the row and each amplicon in the columns}

\item{clusters}{name of the clusters to be analyzed}

\item{amplicons}{name of the amplicons to be analyzed}
}
\value{
matrix
}
\description{
generate a Matrix with the median ploidy for each amplicon suitable for sample-to-sample heatmap and differentially ploidy analysis
}
\examples{
#do not run
}
