% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelevantClones.R
\name{RelevantClones}
\alias{RelevantClones}
\title{RelevantClones}
\usage{
RelevantClones(table, frequency = 15)
}
\arguments{
\item{table}{a data.frame generated by freqclones.py}

\item{frequency}{a cut-off value to consider a clone phase-specific}
}
\value{
a data.frame with phase information
}
\description{
Find clones with an increased frequency in phase2
}
\examples{
#do not run
TACOS::RelevandClones(table, frequency=15)
}
